requirements.txt
asgiref==3.7.2
Django==4.2.4
sqlparse==0.4.4
typing_extensions==4.7.1
djangorestframework==3.14.0
django-rest-authtoken==2.1.4
